requirements.txt
click
rich
google-generativeai
python-dotenv
pwinput
